Move o-admin-edit-event to p-admin-edit-event and remove organism
